api/requirements.txt
fastapi
tortoise-orm
aiomysql
httpx
python-multipart
uvicorn
uvicorn[standard]